refactor: entities folder
