+++ src/modules/identity/authentication/domain/entities/UserCredential.ts
