+++ src/modules/identity/domain/interfaces/repositories/user-credentials-repository.ts
@@ -1,4 +1,4 @@
-import { UserCredential } from "../../../authentication/domain/UserCredential";
+import { UserCredential } from "../../../authentication/domain/entities/UserCredential";
 
 export abstract class UserCredentialsRepository {
   abstract add(user: UserCredential, tx?: unknown): Promise<void>;
